[PR85072, LRA]: Set a limit for considering other reload pseudo preferences
Compilation of test in PR85072 takes a lot of time and memory, e.g. 17
minutes and 23 GB memory on AMD 9900X.  The function in question has one million
program points and one million pseudos. The culprits are
live_reload_and_inheritance_pseudos bitmaps which are used to consider
other reload pseudo preferences when assigning a register to a given
pseudo.  The patch introduces a constraint regarding when those
preferences are considered.  The patch decreases compilation time to
about 10 minutes and memory consumption to about 2GB.

gcc/ChangeLog:

	PR rtl-optimization/85072
	* lra-assigns.cc (init_live_reload_and_inheritance_pseudos):
	Improve calculation of live_reload_and_inheritance_pseudos and set
	a constraint to do this.
	* params.opt
	(lra-max-pseudos-points-log2-considered-for-preferences): New.
	* doc/invoke.texi
	(lra-max-pseudos-points-log2-considered-for-preferences): Document
	it.