arm: avoid invalid shift in arm_canonicalize_comparison [PR122999]
There was UB in arm_canonicalize_comparison if it is called with
both operands of type VOIDmode.  Avoid this by first handling
floating-point types, then returning if we are left with anything
other than an integer mode.  For belt-and-braces also check that
the mode does not require a mask larger than HOST_WIDE_INT.

gcc/ChangeLog:

	PR target/122999
	* config/arm/arm.cc (arm_canonicalize_comparison): Defer
	initializing maxval until we know we are dealing with an
	integer mode.