doc, arc: Clean up ARC option documentation [PR122243]
The ARC front end presently has a large number of options that are
explicitly deprecated, either by the "Warn" option or by being
documented as such in the GCC manual.  The manual text has documented
a long list of obsolete options with a warning that they will be removed
completely in a future release since at least GCC 5, 10+ years ago.
Some of documented options have, in fact, already been deleted.

This patch does *not* delete the remaining obsolete options, but only
marks them as "Undocumented" in the .opt file and removes the
documentation to reduce clutter in the manual.  I've also added missing
index entries for the remaining options to the manual.

gcc/ChangeLog
	PR other/122243
	* config/arc/arc.opt: Mark -mbig-endian, -mlittle-endian,
	-mmixed-code, -mno-mpy, -margonaut, -munalign-prob-threshold=,
	-mannotate-align, -malign-call, -mRcq, -mRcw, -mbbit-peephole,
	-mcompact-casesi, -mq-class, -mexpand-adddi, -mcrc, -mdsp-packa,
	-mdvbf, -mtelephony, -mrtsc, -EB, -EL, -mrla, -mEA, and
	-multcost= as "Undocumented".
	* doc/invoke.texi: Remove documentation for the above options.
	plus -mmac-d16 and -mmac-24 (which were already marked as
	"Undocumented").  Likewise remove documentation for
	-mbarrel_shifter, -mdpfp_compact, -mdpfp_fast, -mdsp_packa,
	-mmac_24, -mmac_d16, -mspfp_compact, and -mspfp_fast, which
	had already been deleted from arc.opt at some point.
	Add index entries for the -mno- forms of remaining options that
	have them.  Document positive forms of -mno-brcc and -mno-dpfp-lrsr.