doc, h8300: Clean up H8/300 option and attribute documentation [PR122243]
Comparing the documentation in invoke.texi with hs8300.opt, I spotted
a few options in the latter that were missing documentation, and added
it.

I wanted to add a link to the "monitor" attribute
referenced in the existing docs for this option, but found that was
also missing, along with docs for the "OS_Task" attribute; so I fixed
those problems while I was at it.

gcc/ChangeLog
	PR other/122243
	* config/h8300/h8300.opt (mexr, mno-exr): Add FIXME re ambiguity
	for -mno-exr semantics.
	* doc/extend.texi (H8/300 Function Attributes): Document
	monitor and OS_Task attributes.
	* doc/invoke.texi (Option Summary) <H8/300 Options}: Add -msx,
	-ms2600, -mquickcall, -mslowbyte.  Combine -mexr and -mno-exr.
	(H8/300 Options): Likewise.  Add @opindex entries for options
	that have negative forms.