doc, pdp11: Clean up PDP-11 documentation [PR122243]
While working on this patch I saw that this target has an -mlra option
that still defaults to off.  Although the LRA support was added in
2018 apparently it wasn't robust enough to enable by default.  We are
supposed to be deleting reload support and all targets that don't use
LRA by default in GCC 16, so this target may be declared obsolete very
soon, but I've made the documentation of other options consistent with
the .opt files and conventions used for other targets anyway in case
either somebody who can build/test for this target switches the
default (see PR target/113947) or the reload removal is postponed.

gcc/ChangeLog
	PR other/122243
	* config/pdp11/pdp11.opt (m40, m45): Add RejectNegative.
	* doc/invoke.texi (Option Summary) <PDP-11 Options>: Remove
	redundant -mno- forms from the list.
	(PDP-11 Options): Fix some markup issues.  Merge documentation
	of positive and negative forms of -mac0.  Index negative forms
	of -msplit, -mlra.