Adjust expected output of new risc-v test
Just a minor update to Dimitar's patch for the RISC-V testcase.

The cfi directives are not emitted for the -elf configurations causing the new
test to fail.  The cfi directives (and associated labels) don't seem relevant
to the test at hand, so this just drops them.

Pushing to the trunk.

	PR rtl-optimization/122675

gcc/testsuite
	* gcc.target/riscv/pr122675-1.c: Adjust expected output.