vect: move the condition down into the loop checking every scalar load [PR122868]
As Richi suggested this moves the check into the loop so we check every load.

I had initially not done this because I figured the loads would be treated as a
group anyway and the group would be valid or not as a whole.  But for invariants
they could be a group, but not all the loads within range of a known bounds.

gcc/ChangeLog:

	PR tree-optimization/122868
	* tree-vect-stmts.cc (vectorizable_load): Move check for invariant loads
	down into the loop.